+++ HISTORY.rst
@@ -9,6 +9,7 @@ Unreleased
 Fixes
 ~~~~~~~~~~~~~~~~~~~~~
 * fix incorrect descriptions returned for ``GET`` requests of job logs and exceptions
+* fix key error when parsing model task defined as literal string [GEOIM-245, GEOIM-247]
 
 1.4.0 (2019-08-06)
 ---------------------
@@ -18,7 +19,7 @@ Changes
 * add error logging in case of request generating an exception (voluntary or by execution problem)
 * change ``gdal`` requirements and imports to avoid ``osgeo`` variant breaking too easily
 * support older ``thelper`` model checkpoint definitions using string parameters instead of JSON [GEOIM-241]
-* update to ``thelper>=0.3.7``
+* update to ``thelper>=0.3.7``, help better support old format for model task definitions [GEOIM-247]
 * disable `Jenkins` online tests execution (download/upload model) as they now require login [GEOIM-180]
 
 Fixes